backend/requirements.txt
python-dotenv==1.0.1
vision-agents[gemini,getstream]==0.3.2
requests==2.31.0